Updates based on changes made today on BridgePF, and eliminating SubpopulationGuid as a string field
