+++ pom.xml
@@ -113,6 +113,7 @@
                 <artifactId>maven-javadoc-plugin</artifactId>
                 <version>2.9.1</version>
                 <configuration>
+                    <show>public</show>
                     <links>
                         <link>http://joda-time.sourceforge.net/apidocs/</link>
                     </links>
