+++ src/main/java/org/sagebionetworks/bridge/sdk/json/SubpopulationGuidDeserializer.java
@@ -4,17 +4,18 @@ import java.io.IOException;
 
 import org.sagebionetworks.bridge.sdk.models.subpopulations.SubpopulationGuid;
 
+import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.KeyDeserializer;
+import com.fasterxml.jackson.databind.JsonDeserializer;
 
-public class SubpopulationGuidKeyDeserializer extends KeyDeserializer {
+public class SubpopulationGuidDeserializer extends JsonDeserializer<SubpopulationGuid> {
 
     @Override
-    public Object deserializeKey(String string, DeserializationContext context)
+    public SubpopulationGuid deserialize(JsonParser jp, DeserializationContext context)
             throws IOException, JsonProcessingException {
-        return new SubpopulationGuid(string);
+        String guid = jp.getText();
+        return new SubpopulationGuid(guid);
     }
 
 }
-
