+++ src/test/java/org/sagebionetworks/bridge/sdk/TestUserHelper.java
@@ -11,7 +11,6 @@ import org.joda.time.LocalDate;
 import org.sagebionetworks.bridge.Tests;
 import org.sagebionetworks.bridge.sdk.exceptions.ConsentRequiredException;
 import org.sagebionetworks.bridge.sdk.models.subpopulations.SubpopulationGuid;
-import org.sagebionetworks.bridge.sdk.models.subpopulations.SubpopulationGuidImpl;
 import org.sagebionetworks.bridge.sdk.models.users.ConsentSignature;
 import org.sagebionetworks.bridge.sdk.models.users.SharingScope;
 import org.sagebionetworks.bridge.sdk.models.users.SignInCredentials;
@@ -58,7 +57,7 @@ public class TestUserHelper {
             return roles;
         }
         public SubpopulationGuid getDefaultSubpopulation() {
-            return new SubpopulationGuidImpl(Tests.TEST_KEY);
+            return new SubpopulationGuid(Tests.TEST_KEY);
         }
         public boolean signOutAndDeleteUser() {
             userSession.signOut();
@@ -109,7 +108,7 @@ public class TestUserHelper {
             userSession = e.getSession();
             if (consent) {
                 ConsentSignature signature = new ConsentSignature("Tester", LocalDate.parse("1970-02-02"), null, null);
-                userSession.getUserClient().consentToResearch(new SubpopulationGuidImpl(Tests.TEST_KEY), signature,
+                userSession.getUserClient().consentToResearch(new SubpopulationGuid(Tests.TEST_KEY), signature,
                         SharingScope.NO_SHARING);
             }
         }
